// game-script/generated/types/BodyCelestialBodyDefinition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BodyAtmosphere } from "./BodyAtmosphere.js";
import type { BodyDynamics } from "./BodyDynamics.js";
import type { BodyPlanetGenConfig } from "./BodyPlanetGenConfig.js";
import type { BodyStarEmission } from "./BodyStarEmission.js";
import type { BodyTerrain } from "./BodyTerrain.js";
import type { BodyWater } from "./BodyWater.js";

export type BodyCelestialBodyDefinition = {
  name: string;
  terrain: BodyTerrain | null;
  water: BodyWater | null;
  atmosphere: BodyAtmosphere | null;
  generatorConfig: BodyPlanetGenConfig | null;
  dynamics: BodyDynamics;
  starEmission: BodyStarEmission | null;
};